+++ app/Http/Controllers/TipoUsuarioController.php
@@ -6,22 +6,27 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 use App\Models\User;
+use App\Http\Middleware\AdminMiddleware;
+
 
 class TipoUsuarioController extends Controller
 {
     //
-    public function tipoUsuario()
+   
+       public function tipoUsuario()
     {
-        $auth = Auth::user();
-        if($auth->tipo_usuario !== 'admin'){
-        return redirect()->route('minha-conta');
-        } else{
+       
             return Inertia::render('Dashboard');
-        }
+        
  
     }
     public function minhaConta()
     {
         return 'minha conta';
+    }
+
+    public function todoProduto()
+    {
+        return Inertia::render("TodoProduto");
     }
 }