+++ resources/js/Pages/TodoProduto.jsx
@@ -1,0 +1,13 @@
+import AuthenticatedLayout from '../Layouts/AuthenticatedLayout';
+import { Head } from '@inertiajs/react';
+
+export default function TodoProduto()
+{
+  return (
+    <>
+    <AuthenticatedLayout>
+      ola
+    </AuthenticatedLayout>
+    </>
+  )
+}