+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 
 # [1.0.0] - 2025-02-10
 - Tabela de checklist com desconto e botãoo para aplicar o desconto
+# [1.0.0] - 2025-02-12
+- Modificação no Middleware adicionado class admin 
+- Modificação no codigo do arquivo controller (TipoUsuario)